Triggers are now importable from covalent directly
