--- covalent_dispatcher/_core/trigger_testing.txt
@@ -1,8 +1,0 @@
-dsadasddsadasdasdasd
-sdasadasd
-sdasdasdasd
-sdasdasdasdsadas
-sadsaddsad
-sdasasad
-dsa
-fdsfdsaddssadasdasd
